(hopefully) finished frontend

+++ backend.py
@@ -1,20 +1,11 @@
-from proxmoxer import ProxmoxAPI
-import os
-from dotenv import load_dotenv
 import time
-import sqlite3
-
-conn = sqlite3.connect("database.db")
-cursor = conn.cursor()
-
-load_dotenv()
-
-ROOTPASS = os.environ["ROOTPASS"]
-CONTROLLERADDR = os.environ["CONTROLLERADDR"]
-
-treknet = ProxmoxAPI(CONTROLLERADDR, user="root@pam", password=ROOTPASS, verify_ssl=False)
+import proxmoxer
 
 def cloneVM(dbcurs, api, vm, userid):
+    dbcurs.execute(f"SELECT * FROM USERDATA WHERE USERID = '{userid}';")
+    check = dbcurs.fetchone()
+    if check != None:
+        return False
     dbcurs.execute("SELECT MAX(VMID) FROM USERDATA;")
     try:
         newid = dbcurs.fetchone()[0] + 1
@@ -24,7 +15,7 @@ def cloneVM(dbcurs, api, vm, userid):
     try:
         vncport = dbcurs.fetchone()[0] + 1
     except TypeError:
-        vncport = 77
+        vncport = 1
     newNode = min(scoreNodes(api), key = scoreNodes(api).get)
     if vm == "ubuntu":
         vmid = 104
@@ -43,13 +34,44 @@ def cloneVM(dbcurs, api, vm, userid):
     time.sleep(3)
     api.nodes(newNode).qemu(newid).config().post(args=f"-vnc 0.0.0.0:{vncport}")
     dbcurs.execute(f"INSERT INTO USERDATA VALUES ('{userid}', {newid}, {vncport}, '{newNode}');")
+    return [newid, vncport, newNode]
+
+def powerVM(dbcurs, api, userid, action):
+    dbcurs.execute(f"SELECT * FROM USERDATA WHERE USERID = '{userid}';")
+    vm = dbcurs.fetchone()
+    if vm == None:
+        return False
+    match action:
+        case "shutdown":
+            api.nodes(vm[3]).qemu(vm[1]).status().shutdown().post(node=vm[3], vmid=vm[1])
+        case "start":
+            api.nodes(vm[3]).qemu(vm[1]).status().start().post(node=vm[3], vmid=vm[1])
+        case "reboot":
+            api.nodes(vm[3]).qemu(vm[1]).status().reboot().post(node=vm[3], vmid=vm[1])
+        case "stop":
+            api.nodes(vm[3]).qemu(vm[1]).status().stop().post(node=vm[3], vmid=vm[1])
+        case "reset":
+            api.nodes(vm[3]).qemu(vm[1]).status().reset().post(node=vm[3], vmid=vm[1])
+        case _:
+            return
 
 def delVM(dbcurs, api, userid):
     dbcurs.execute(f"SELECT * FROM USERDATA WHERE USERID = '{userid}';")
     vm = dbcurs.fetchone()
-    print(vm)
-    api.nodes(vm[3]).qemu(vm[1]).delete()
-    dbcurs.execute(f"DELETE FROM USERDATA WHERE USERID = '{userid}';")
+    if vm == None:
+        return False
+    try:
+        api.nodes(vm[3]).qemu(vm[1]).delete()
+        dbcurs.execute(f"DELETE FROM USERDATA WHERE USERID = '{userid}';")
+    except proxmoxer.core.ResourceException:
+        return 10
+
+def vmstat(dbcurs, userid):
+    dbcurs.execute(f"SELECT * FROM USERDATA WHERE USERID = '{userid}';")
+    vm = dbcurs.fetchone()
+    if vm == None:
+        return False
+    return [vm[1], vm[2], vm[3]]
 
 def scoreNodes(api):
     nodes = {}
@@ -57,13 +79,3 @@ def scoreNodes(api):
         if node["status"] == "online":
             nodes[node["node"]] = node["cpu"] * node["mem"]
     return nodes
-
-#delVM(cursor, treknet, "test1")
-#delVM(cursor, treknet, "test2")
-#delVM(cursor, treknet, "test3")
-#cloneVM(cursor, treknet, 104, "test1")
-#cloneVM(cursor, treknet, 104, "test2")
-#cloneVM(cursor, treknet, 104, "test3")
-
-conn.commit()
-conn.close()
